[DON'T MERGE] [RFC] Optimize access to GlobalConst
The change ensures the global constants can be accessed without
conditional on whether they were initialized. Before this the constants
were accessed via function that would check if the initialization was
complete. If it wasn't, then the initialization was performed before
the function returns. With this change, the initialization happens
early when the DSO (or static binary) is loaded into the process - via
the init_array. This way the conditional can be removed everywhere the
constants are accessed during normal execution.

Additionally to solving the issue with initialization order across
libraries, the change ensures the global constants can be accessed
directly via offset instead of being accessed via the Global Offset
Table.

This patch is an RFC on the alternative approach to ensure static
initialization order is correct for the gPageSize and dependent
global constants however possibly it can be seen as providing an
infrastructure for all global constant values independently of page
size agnosticism support.

+++ runtime/gc/collector/mark_compact.cc
@@ -1955,7 +1955,7 @@ void MarkCompact::MapProcessedPages(uint8_t* to_space_start,
   // so in a single ioctl. This helps avoid the overhead of invoking syscall
   // several times and also maps the already-processed pages, avoiding
   // unnecessary faults on them.
-  size_t length = kFirstPageMapping ? gPageSize : 0;
+  size_t length = kFirstPageMapping ? gPageSize : (size_t) 0;
   if (kFirstPageMapping) {
     arr_idx++;
   }
@@ -4433,7 +4433,7 @@ void MarkCompact::FinishPhase() {
   // userfault, which maps a special zero-page.
   if (use_uffd_sigbus_ || !minor_fault_initialized_ || !shadow_to_space_map_.IsValid() ||
       shadow_to_space_map_.Size() < (moving_first_objs_count_ + black_page_count_) * gPageSize) {
-    size_t adjustment = use_uffd_sigbus_ ? 0 : gPageSize;
+    size_t adjustment = use_uffd_sigbus_ ? (size_t) 0 : gPageSize;
     ZeroAndReleaseMemory(compaction_buffers_map_.Begin() + adjustment,
                          compaction_buffers_map_.Size() - adjustment);
   } else if (shadow_to_space_map_.Size() == bump_pointer_space_->Capacity()) {
